+++ src/domain/combat/index.ts
@@ -2,6 +2,7 @@ export * from "./combatActions";
 export * from "./combatBalance";
 export * from "./combatEngine";
 export * from "./combatFlavor";
+export * from "./doppelgangerLines";
 export * from "./combatProbe";
 export * from "./combatState";
 export * from "./monsterCombatStats";